refactor(bridge): rewrite PrSyncService to depend on repositories
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
